requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic>=2.10.0
pydantic-settings>=2.7.0
httpx==0.27.0
python-dotenv==1.0.1
email-validator==2.1.1
Pillow
aiofiles==23.2.1